requirements.txt
# Consolidated BrandGuard Pipeline Dependencies

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0

# Machine Learning and AI
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
transformers>=4.20.0
scikit-learn>=1.0.0

# OCR and Text Processing
google-cloud-vision>=3.7.2
pytesseract>=0.3.10
nltk>=3.7
textblob>=0.17.1
pyspellchecker>=0.7.0

# Image and Document Processing
PyMuPDF>=1.19.0
pdf2image>=1.17.0
beautifulsoup4>=4.10.0

# Configuration and Validation
PyYAML>=6.0
pydantic>=1.9.0

# Web Framework and Utilities
Flask>=2.0.0
Flask-CORS>=3.0.10
Werkzeug>=2.0.0
requests>=2.25.0

# FastAPI and JSON API support
fastapi[all]>=0.111.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # file uploads
aiofiles>=23.0.0  # static file serving

# Visualization and Utilities
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pathlib2>=2.3.0
python-dateutil>=2.8.0
scikit-image==0.24.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910


